+++ src/containers/Form/DataAccessReports/__tests__/index.test.js
@@ -11,7 +11,7 @@ import {
   getSelectedEntityId,
   isInherited,
   isCreating,
-  userHasUpdatePermission,
+  userHasUpdatePermission
 } from '../../../../redux/modules/entities/selectors';
 import { selectFormInitialValues } from '../../../../redux/modules/form/selectors';
 import { selectDashboards } from '../../../../redux/modules/entities/dashboards/selectors';
@@ -30,9 +30,7 @@ selectFormInitialValues.mockImplementation(() => ({ active: true }));
 
 jest.mock('../../../../redux/modules/entities/dashboards/selectors');
 selectDashboards.mockImplementation(() => [{ name: 'mockName' }]);
-jest.mock(
-  '../../../../redux/modules/entities/historicalReportFolders/selectors'
-);
+jest.mock('../../../../redux/modules/entities/historicalReportFolders/selectors');
 selectHistoricalReportFolders.mockImplementation(() => [{ name: 'mockName' }]);
 jest.mock('../../../../redux/modules/entities/users/selectors');
 selectNonDisabledUsers.mockImplementation(() => [{ name: 'mockName' }]);
@@ -40,11 +38,7 @@ selectNonDisabledUsers.mockImplementation(() => [{ name: 'mockName' }]);
 describe('DataAccessReports Renders', () => {
   it('renders', () => {
     const store = createStore(state => state);
-    expect(
-      shallow(
-        <DataAccessReportsForm store={store}>Child</DataAccessReportsForm>
-      )
-    ).toMatchSnapshot();
+    expect(shallow(<DataAccessReportsForm store={store}>Child</DataAccessReportsForm>)).toMatchSnapshot();
   });
 });